Introduce Type base class for ComplexType/UniqueType hierarchy
Ports castwide/solargraph#1162 onto current master. Extracts the
common type-data methods previously duplicated between ComplexType
and ComplexType::UniqueType (via the TypeMethods mixin) into a new
Solargraph::Type base class that both now inherit from.

The original PR's branch had diverged from master across ~1300
commits; most of its 269-file diff was master having independently
superseded it (inline RBS support, YARD macros, literal-type-inference
rollback, pin caching rework, etc.). This port keeps only the type
hierarchy refactor itself and re-resolves everything else against
current master.
