+++ lib/solargraph/complex_type.rb
@@ -3,18 +3,20 @@
 module Solargraph
   # A container for type data based on YARD type tags.
   #
-  class ComplexType
+  class ComplexType < Type
     GENERIC_TAG_NAME = 'generic'
+
     # @!parse
     #   include TypeMethods
     include Equality
 
     autoload :Conformance, 'solargraph/complex_type/conformance'
-    autoload :TypeMethods, 'solargraph/complex_type/type_methods'
     autoload :UniqueType,  'solargraph/complex_type/unique_type'
 
     # @param types [Array<UniqueType, ComplexType>]
     def initialize types = [UniqueType::UNDEFINED]
+      super()
+
       # @todo @items here should not need an annotation
       # @type [Array<UniqueType>]
       items = types.flat_map(&:items).uniq(&:to_s)
@@ -44,7 +46,7 @@ module Solargraph
     end
 
     # @param generics_to_resolve [Enumerable<String>]]
-    # @param context_type [ComplexType, ComplexType::UniqueType, nil]
+    # @param context_type [Type, nil]
     # @param resolved_generic_values [Hash{String => ComplexType}] Added to as types are encountered or resolved
     # @return [self]
     def resolve_generics_from_context generics_to_resolve, context_type, resolved_generic_values: {}
@@ -68,7 +70,7 @@ module Solargraph
        (@items.length > 1 ? ')' : ''))
     end
 
-    # @param dst [ComplexType, ComplexType::UniqueType]
+    # @param dst [Type]
     # @return [ComplexType]
     def self_to_type dst
       object_type_dst = dst.reduce_class_type
@@ -163,10 +165,15 @@ module Solargraph
       super
     end
 
+    def undefined?
+      @items.all?(&:undefined?)
+    end
+
     # @param name [Symbol]
     # @param include_private [Boolean]
     def respond_to_missing? name, include_private = false
-      TypeMethods.public_method_defined?(name) || super
+      return false if @items.first.nil?
+      @items.first.respond_to?(name) || super
     end
 
     def to_s
@@ -199,7 +206,7 @@ module Solargraph
     end
 
     # @param api_map [ApiMap]
-    # @param expected [ComplexType, ComplexType::UniqueType]
+    # @param expected [Type]
     # @param situation [:method_call, :return_type, :assignment]
     # @param rules [Array<:allow_subtype_skew, :allow_empty_params, :allow_reverse_match, :allow_any_match, :allow_undefined, :allow_unresolved_generic, :allow_unmatched_interface>]
     #
@@ -378,7 +385,7 @@ module Solargraph
 
     # @see https://en.wikipedia.org/wiki/Intersection_type
     #
-    # @param intersection_type [ComplexType, ComplexType::UniqueType, nil]
+    # @param intersection_type [Type, nil]
     # @param api_map [ApiMap]
     # @return [self, ComplexType::UniqueType]
     def intersect_with intersection_type, api_map
