+++ lib/solargraph/pin/common.rb
@@ -10,7 +10,7 @@ module Solargraph
       #   @abstract
       #   @return [void]
       # @type @closure [Pin::Closure, nil]
-      # @type @binder [ComplexType, ComplexType::UniqueType, nil]
+      # @type @binder [Type, nil]
 
       # @todo Missed nil violation
       # @return [Location, nil]
@@ -44,7 +44,7 @@ module Solargraph
         @return_type ||= ComplexType::UNDEFINED
       end
 
-      # @return [ComplexType, ComplexType::UniqueType]
+      # @return [Type]
       def context
         # Get the static context from the nearest namespace
         @context ||= find_context
@@ -56,7 +56,7 @@ module Solargraph
         context.namespace.to_s
       end
 
-      # @return [ComplexType, ComplexType::UniqueType]
+      # @return [Type]
       # @sg-ignore https://github.com/castwide/solargraph/pull/1100
       def binder
         @binder || context
